pyproject.toml
[tool.poetry]
name = "mycalendar"
version = "0.1.0"
description = "Calendar"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.5.3"
nose2 = "^0.9.2"
pytruth = "^1.1.0"
blinker = "^1.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
